Add io module to models
