+++ calkit/models/__init__.py
@@ -1,2 +1,3 @@
 from .core import *  # noqa: F403, I001
 from . import pipeline  # noqa: F401, F403
+from . import io  # noqa: F401, F403
