+++ calkit/models/io.py
@@ -1,0 +1,49 @@
+"""Models for inputs and outputs."""
+
+from typing import Literal
+
+from pydantic import BaseModel, ConfigDict
+
+
+class Input(BaseModel):
+    kind: Literal["path", "python-object", "file-segment", "database-table"]
+
+
+class PathInput(Input):
+    kind: Literal["path"]
+    path: str
+
+
+class PythonObjectInput(Input):
+    kind: Literal["python-object"]
+    module: str
+    object_name: str
+
+
+class FileSegmentInput(Input):
+    kind: Literal["file-segment"]
+    path: str
+    start_line: int
+    end_line: int
+
+
+class DatabaseTableInput(Input):
+    kind: Literal["database-table"]
+    database_uri: str
+    database_name: str | None = None
+    table_name: str
+
+
+class PathOutput(BaseModel):
+    path: str
+    storage: Literal["git", "dvc"] | None = "dvc"
+    delete_before_run: bool = True
+    # Do not allow extra keys
+    model_config = ConfigDict(extra="forbid")
+
+
+class DatabaseTableOutput(BaseModel):
+    kind: Literal["database-table"]
+    uri: str
+    database_name: str | None = None
+    table_name: str
