+++ calkit/models/pipeline.py
@@ -8,6 +8,7 @@ from typing import Literal
 from pydantic import BaseModel, ConfigDict, Discriminator
 from typing_extensions import Annotated
 
+from calkit.models.io import PathOutput
 from calkit.models.iteration import (
     ParameterIteration,
     ParametersType,
@@ -19,52 +20,8 @@ from calkit.notebooks import (
 )
 
 
-class Input(BaseModel):
-    kind: Literal["path", "python-object", "file-segment", "database-table"]
-
-
-class PathInput(Input):
-    kind: Literal["path"]
-    path: str
-
-
-class PythonObjectInput(Input):
-    kind: Literal["python-object"]
-    module: str
-    object_name: str
-
-
-class FileSegmentInput(Input):
-    kind: Literal["file-segment"]
-    path: str
-    start_line: int
-    end_line: int
-
-
-class DatabaseTableInput(Input):
-    kind: Literal["database-table"]
-    database_uri: str
-    database_name: str | None = None
-    table_name: str
-
-
 class InputsFromStageOutputs(BaseModel):
     from_stage_outputs: str
-
-
-class PathOutput(BaseModel):
-    path: str
-    storage: Literal["git", "dvc"] | None = "dvc"
-    delete_before_run: bool = True
-    # Do not allow extra keys
-    model_config = ConfigDict(extra="forbid")
-
-
-class DatabaseTableOutput(BaseModel):
-    kind: Literal["database-table"]
-    uri: str
-    database_name: str | None = None
-    table_name: str
 
 
 class StageIteration(BaseModel):
